+++ test/unit/mongo_db_admin/run_dbcc.py
@@ -41,10 +41,10 @@ class Mongo(object):
     Description:  Class stub holder for mongo_class.DB class.
 
     Methods:
-        __init__ -> Class initialization.
-        validate_tbl -> Stub holder for mongo_class.DB.validate_tbl method.
-        chg_db -> Stub holder for mongo_class.DB.chg_db method.
-        get_tbl_list -> Stub holder for mongo_class.DB.get_tbl_list method.
+        __init__
+        validate_tbl
+        chg_db
+        get_tbl_list
 
     """
 
@@ -120,9 +120,9 @@ class UnitTest(unittest.TestCase):
     Description:  Class which is a representation of a unit testing.
 
     Methods:
-        setUp -> Initialize testing environment.
-        test_tbl_list -> Test with table list.
-        test_default -> Test with default arguments.
+        setUp
+        test_tbl_list
+        test_default
 
     """