+++ CHANGELOG.md
@@ -4,7 +4,28 @@ All notable changes to this project will be documented in this file.
 The format is based on "Keep a Changelog".  This project adheres to Semantic Versioning.
 
 
+## [2.4.0] - 2021-09-09
+- Updated to work in Mongo 4.2.14 environment.
+- Updated to work in a SSL environment.
+- Added ability to override the default mail command and use mailx.
+
+### Added
+- process_dbs_tbls:  Process a list of databases and tables.
+- process_mail:  Add data to mail instance and send mail.
+
+### Changed
+- process_request:  Added auth_mech to mongo_class instance call.
+- main:  Add -u option to the conditional required list.
+- config/mongo.py.TEMPLATE:  Added SSL connection entries.
+- process_request:  Replaced process databases and tables with call to process_dbs_tbls and combined a number of if and else statements into a single statement.
+- status:  Replaced email section with call to process_mail.
+- run_compact:  Combined if and else into a single statement.
+- Documentation updates.
+
+
 ## [2.3.2] - 2020-12-29
+- Verified to work with pymongo v3.8.0.
+- Updated to be used in FIPS 140-2 environment.
 
 ### Fixed
 - process_request:  When using the -t option, do not process a database if no tables are present.
@@ -34,7 +55,7 @@ The format is based on "Keep a Changelog".  This project adheres to Semantic Ver
 - run_program:  Removed sys.exit() call.
 - status:  Refactored the email code section to reduce complexity.
 - config/mongo.py.TEMPLATE:  Set the replica set variables to None.
-- status:  Changed JSON key values to CamelCase.
+- status:  Changed JSON key values to PascalCase.
 - rotate:  Added option compression call when rotating log files to -n option directory.
 - main:  Added program lock functionality to program.